requirements.txt
urllib3<2.0
youtube-transcript-api
pytube
openai
langchain
python-dotenv
beautifulsoup4
requests
google-api-python-client
google-auth-oauthlib
google-auth
google-auth-httplib2
lxml
tiktoken